fix: the function file_get_contents could return false
prevent to go ahead if got error to read the temp file

+++ src/Sign/JSignService.php
@@ -190,6 +190,9 @@ class JSignService
                 REPACK_COMMAND
             );
             $certificateRepacked = file_get_contents($tempEncriptedRepacked);
+            if ($certificateRepacked === false) {
+                return [];
+            }
             $params->setCertificate($certificateRepacked);
             unlink($tempPassword);
             unlink($tempEncriptedOriginal);
